Add files via upload
Fixed a MAJOR ERROR that affected all approximation methods. The condition x < end is unreliable when dealing with floating-point values, so Math.abs(x - end) > EPSILON is used instead. For the MMS() and RMS() methods, the "end" parameter in the call to LRS() was incorrect (it needed to be incremented by deltaX/2 and deltaX, respectively).